fixed the homepage now the messages arent showing on there and the extra link isnt floating around
