+++ Stranger-Things/package.json
@@ -1,8 +1,9 @@
 {
   "name": "stranger-things",
-  "private": true,
+  "private": false,
   "version": "0.0.0",
   "type": "module",
+  
   "scripts": {
     "dev": "vite",
     "build": "vite build",
