+++ Stranger-Things/src/components/HomePage.jsx
@@ -1,32 +1,19 @@
-//HomePage.jsx
-
-import React from "react";
+//Homepage should only have Nav tabs and welcome message
 import Navbar from "./NavBar";
-import { useNavigate } from "react-router-dom";
-import "./HomePage.css";
-
-const Home = ({ token, setToken }) => {
-  const navigate = useNavigate();
-
-  const handleLogOut = () => {
-    setToken(null);
-    navigate("/login");
-  };
-
+export const Home = () => {
   return (
     <div className="home">
-      <Navbar token={token} handleLogOut={handleLogOut} />
-      <div>
-        <button onClick={handleLogOut}>Log Out</button>
-        
-        {/* Updated content */}
-        <h1>Welcome to Our Message Board</h1>
-        <p>Engage in discussions, share ideas, and connect with others.</p>
-        <p>Discover a diverse range of topics and perspectives.</p>
-        <p>Join the conversation and start connecting!</p>
+      <Navbar />
+        <div>
+      <h1>Welcome to MCJ Connections!</h1>
       </div>
     </div>
   );
 };
 
+
 export default Home;
+
+        
+      
+
