+++ Stranger-Things/src/components/Messages.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+export const Messages = () => {
+    
+
+  return (
+      <div id='messages'>
+      <h1>Welcome to Our Message Board</h1>
+      <p>Engage in discussions, share ideas, and connect with others.</p>
+      <p>Discover a diverse range of topics and perspectives.</p>
+      <p>Join the conversation and start connecting!</p>
+    </div>
+  
+  )
+}
